refactor: move the functions from dump.go to sql.go
